AWB-162: Fix a void target call that occurred when re-selecting a specialised archetype in the repository tree, where the specialisation parent is missing.
The problem was that ARCH_REP_ARCHETYPE.is_specialised was being used to protect against calling a void specialisation parent, presumably on the assumption that is_specialised would return false if the specialisation parent was void. But is_specialised merely checks the specialisation status of the archetype id: it says nothing about whether the specialisation parent is attached. Also, the protection was using "and" rather than "and then".