Add a test that loads the test archetype repository and then compiles all of the validation test archetypes.
This is just to get a feel for how to do this in the EiffelStudio testing framework. Currently it just looks at the text in the message billboard, which is very crude. We will want to check specific validity codes and report exactly which archetypes are not doing what we expect.

Also, it tries to print the billboard contents, but this does not seem to work when run within EiffelStudio with the adl_workbench project. If we want to print details about test failures, it looks like we would probably need a console project. That's okay, because we would need to do that anyway in order to run the tests within CruiseControl.

For now, to run the test as it is within EiffelStudio, it's easier to see what's going on by clicking the Testing tool's "debug" button rather than the "run" button.