AWB-167: Eiffel Software is planning to remove the implicit conversions from STRING_32 to STRING_8 in EiffelStudio 6.5. See http://tech.groups.yahoo.com/group/eiffel_software/message/15487
I tested the effect of the conversion removal in our code by editing the STRING_32 class and removing `as_string_8' from the list of conversion routines. Then I compiled adl_workbench and adl_parser (with the no_precompile target) and fixed the errors, either by adding explicit calls to 'as_string_8' or else by changing the type of the target entity to STRING_32.