Add postconditions and unit test assertions demonstrating that a new minimal differential archetype was created in an invalid state.
The bug was introduced in revision 338. At that time, there was only one class each for ARCHETYPE and ARCHETYPE_ONTOLOGY: the differential and flat descendant classes had not been written yet. Prior to then, when ARCHETYPE.make_minimal created the archetype's ontology, it passed in the concept term object with two items, "text" and "description" (both with the value "unknown"). Revision 338 changed this. A major advance in that revision was to create the ontology at the correct specialisation depth. To this end, revision 338 now passed in the specialisation depth to ARCHETYPE_ONTOLOGY.make_empty, instead of the concept term; and so the newly created ontology object was now responsible for creating its own concept term, at the correct depth. Unfortunately, the need to create the concept term's "text" and "description" items seems to have been forgotten, causing the bug.

Because Archetype Editor has been using an ADL Parser based on code from before revision 338, and because ADL Workbench does not actually create new archetypes (it only works with existing archetypes), no live code has actually ever called DIFFERENTIAL_ARCHETYPE.make_minimal; and so this problem has gone unnoticed for nearly three years. It has come to light now while upgrading Archetype Editor to use the the current ADL Parser source code. One of Archetype Editor's main functions is to create a new archetype, and so it was getting an error when trying to save the archetype. In order to generate the archetype's MD5 hash, the canonical XML serialisation is generated. The archetype was failing XML schema validation because the concept term had no items.

The fix is in DIFFERENTIAL_ARCHETYPE_ONTOLOGY.make_empty. As ARCHETYPE.make_minimal used to do prior to revision 338, it now adds the "text" and "description" items to the concept term.

(Note that I initially attempted to fix this by changing ARCHETYPE_TERM.make to add the two items. This did not work, because ARCHETYPE_TERM.make is called from several other places which rely on it leaving the items empty. This violated the Key_valid precondition of ARCHETYPE_TERM.add_item: a_key /= Void and then not has_key(a_key). For example, when ARCHETYPE_ONTOLOGY.populate_term_defs tries to add the "text" and "description" items, the precondition would fail if the items were already there. Avoiding this would involve either clearing the items immediately before populating them, which seems like a hack; or else relaxing the Key_valid precondition, which seems contrary to the spirit of ARCHETYPE_TERM.add_item. I chose instead to add a postcondition to ARCHETYPE_TERM.make, to make its behaviour explicit, limiting my implementation changes to DIFFERENTIAL_ARCHETYPE_ONTOLOGY.make_empty.)