After the user closes the repository dialog, there were two problems: * The main window did not have focus. * While the repository was being reloaded, there was a messy remnant of the dialog still visible in front of the main window. * If the user did something like typing Ctrl+Space,X to maximise the main window, the message actually caused the closed dialog to reopen temporarily.
These problems all had one cause. The underlying window resource had not been destroyed, because garbage collection had not kicked in to dispose of it yet. All three problems are fixed by explicitly destroying the dialog as soon as we are done with it.

(I'm not sure why other dialogs are not exhibiting the same problem. We may need to destroy them explicitly also.)