Fix a usability problem on Windows (but not on GTK) when the user is doing a search in the archetype id combo box. Pressing the <Down> key to select an id in the combo box's drop-down list triggered the select action immediately. This caused an archetype to be selected in the explorer tree, and the user was unable to select the desired id via the keyboard. The drop-down list also remained open, but empty, leaving the screen in an ugly state.
This is fixed in the select action handler. If the drop-down list is still open, then nothing is done. This prevents the wrong archetype from being selected, and the user is able to keep hitting the <Down> key to the desired id. If, on the other hand, the drop-down list is closed, then the archetype is selected, allowing selection via the mouse to work properly, as before.

In addition to this, a fix is needed in the <Enter> key handler. If the drop-down list is open when <Enter> is pressed, then the list is closed, fixing the ugly cosmetic problem. Then the archetype is selected in the explorer tree.

(These fixes are applied only if the platform is Windows. Not only are the fixes not needed on GTK platforms, which were already behaving properly, but the fixes for Windows would actually cause other problems on GTK, preventing the archetype from being selected until after <Enter> had been hit a second time.)