Update the SCons Eiffel builder to work properly with EiffelStudio 6.5 and higher. In EiffelStudio 6.4 and earlier, if the -project_path flag was not supplied then the EIFGENs directory was created under the ECF file. EiffelStudio 6.5 changed this so that it would be created under the current working directory. See http://eiffelstudio.origo.ethz.ch/node/268 for a discussion of why.
This broke the builder's emitter, which was now returning the target in the wrong location, so that subsequent build steps could not find the built files. The emitter now uses the current working directory, fixing the problem for EiffelStudio 6.5 and above. This means that it is now broken for older versions of EiffelStudio, but presumably no one will care about that.

This brought another problem to the fore. On Windows, the .rc file, and any icon resources mentioned in the .rc file, cannot be found by the Eiffel compiler if the EIFGENs path is not under the ECF. This has always been a problem if the -project_path flag was used, and these files have always had to be copied manually to the relevant directory. The builder now copies the .rc file automatically immediately before running the Eiffel compiler. While doing so, it substitutes the path to any icon resources so that the icons can be found below the ECF directory. When compilation has finished, it removes the copy of the .rc file.