AWB-167: Update our override of the .NET version of INTERNAL for EiffelStudio 6.6. (This applies a fix for type conformance of generics).
Note that a small change was also required to adl_parser.ecf. Without this change, there was a peculiar compiler error in the base library's TYPE class, complaining that the actual argument passed to a routine in INTERNAL did not conform to the formal argument. The argument was of a class declared in EiffelSoftware.Runtime.dll. The problem was that there are two copies of this DLL, in EiffelStudio's bin and lib directories, and that adl_parser.ecf was referring to the copy in "bin" but base.ecf is referring to the copy in "lib". The fix is to make adl_parser.ecf also refer to "lib".