Add a Subversion external to the rm_schemas directory in the knowledge2 repository's TRUNK. This will allow us to delete the copies of it currently under each application area.
Currently it is fixed to revision 21 of the knowledge2 repository; although fixing to a specific revision is a bit of a nuisance, requiring a manual change to the revision specified by the external reference, unless you do this there would be no way to know which revision of the external was meant to work with any specific revision or branch of the current repository.

Update the SConstruct and NSIS scripts to point to the rm_schemas directory's new location.