Fix a couple of more compilation errors in classes that are in the universe of classes for the dadl_test project, but not compiled into the system. EiffelStudio 6.7 beta generates compiler errors for these classes because, if the testing library is included, everything is now compiled.
The dadl_test project still does not compile with EiffelStudio 6.7 beta. It appears to be impossible to do so without reorganising other clusters, because there are dependencies on classes in clusters that also contain other classes that have the same name as classes in dadl_test, e.g. APP_ROOT appears in two clusters.