Add a unit test demonstrating that rebuilding a repository that had already been built violated a precondition that there must not be more compiled archetypes than the total number of archetypes in the repository. The problem occurred, in fact, even if only part of the repository had previously been built. It could also occur if rebuilding only a sub-tree.
Fix the problem by decrementing the count of compiled archetypes when resetting the last compilation timestamp on each archetype prior to rebuilding. I chose this way of fixing it, rather than simply resetting the count to zero, because the latter would work only when rebuilding the whole repository; it wouldn't work for a sub-tree.

TODO: Similar problems probably exist for other statistical counters.