Revision 1404 ensured that reselection of the same node in the archetype or template tree caused the widgets to be repopulated from that archetype or template. It did so by firing the select_actions event on the relevant tree widget whenever it got focus. Prior to that fix, the widgets sometimes displayed the contents of an archetype even though a template was currently selected, and vice versa.
Unfortunately, this caused a new problem. The widgets were repopulated whenever the archetype or template tree got focus, even if the other tree had not had focus in the meantime. This could happen, for example, if the user switched back to ADL Workbench from another application. This caused visible activity on the screen, which was disconcerting given that the user had not selected a different item.

The fix is to suppress the behaviour if the tree item selected is the same as the current archetype directory's selected item.