Fix a precondition violation in ARCHETYPE_ONTOLOGY.remove_term_binding, which was failing when called from remove_term_definition. The failing precondition asserts that terminologies_available has the given terminology. It was failing despite the fact that it actually did have the terminology, because terminologies_available was doing reference comparisons. Although the creation routines ensure that terminologies_available will do object comparisons, when the ontology is instantiated via DT_OBJECT_CONVERTER this does not happen. Although revision 1423 added ARCHETYPE_ONTOLOGY.finalise_dt in order to ensure that various things would be cleaned up, it missed this detail.
The problem is now fixed in ARCHETYPE_ONTOLOGY.finalise_dt by setting object comparison on terminologies_available. This is a bit awkward because it may already have been populated, so we have to make a copy of its contents, then clear it out, then set the flag and, finally, copy the contents back into it.

I've applied a similar fix for the term_codes and constraint_codes collections. Although I haven't seen an actual assertion failure for these, it looks to me that it might be possible that they could have the same problem. Perhaps this is unnecessary for those two.