Working interim update of dADL (what was called 'ADL2') archetypes. At this point, compiled archetypes are being successfully round-tripped through a pure dADL serialisation. This is being demonstrated in new tests in the 'Test' page, and a new 'diff' button that allows ADL1.5 text created by the normal ADL parser/serialiser to be compared to ADL text created by serialising an in-memory AOM archetype structure to dADL, via the P_XXX classes, and back again, and then back out through the ADL serialiser. The only remaining difference is that the re-materialised archetypes don't have their parent links set, nor the parent link of the ontology.
The next stage is to integrate this round-tripping more closely into the compiler i.e. ARCHETYPE_COMPILER, ARCH_REP_ARCHETYPE, ARCHETYPE_VALIDATOR, ARCHETYPE_FLATTENER classes, so that compilation state is preserved across sessions, and the AWB can be started and instantly display the compilation state of the last session, and also, for each archetype, where appropriate, to materialise it in memory from the dADL compiled form rather than reparsing and validating. This is somewhat tricky, since it reqires some re-arrangement of the compiler state machine (see ARCH_REP_ARCHETYPE) and also use of the serialised flat from the parent of any specialised archetype.

The dADL serialisation capability serves at least 2 purposes: smarter compilation, as per the above, and also as a basis for XML OPT output, based on the P_XXX classes (remembering that the latter provide a somewhat 'clean' post-parse serialisation of an archetype suitable for persistence). The XML generation can be done from the DT structure of a P_ARCHETYPE, via a new XML serialiser (i.e. a sibling of DADL_SYNTAX_SERIALISER).

A third obvious use for dADL (or the equivalent XML-ised) validated archetypes & templates is in operational systems, where no source level parsing should take place.