This set of changes introduces a basic rule-based framework for driving XML serialisation from a Data Tree (DT). The sample_xml_rules.cfg shows what the rules file may look like; current contents designed to allow archetype serialisation based on conversion of AOM => P_AOM structures, then to DT, then to serialised output.
The xml_rules.cfg file, if found in the use app config area, is materialised into an instance of XML_RULES which is then used by DT_XML_SERIALISER. If no file is found, a reasonable set of basic defaults are created in memory.

Some refactoring was made during this change; the 'rm_type_name' and 'rm_attr_name' attributes in the DT_X classes are renamed to im_type_name and im_attr_name, to reflect the generic nature of the DT structure - everything is just an 'information model' to it - it knows nothing of particular 'reference models', openEHR or anything else special.

Serialisation of DT_PRIMITIVE_OBJECT_LIST data is now correct - a series of single tagged items are now generated, as would be expected in normal XML.

The next steps include reviewing and upgrading the AOM 1.4 published XSD, and then modifying the XML output to comply to that.