Reasonable (but rough) layout capture, using internal docking_manager facility. It looks for a file ui_config.cfg in the user config area for the app; if it doesn't find that, it looks for a default_config_ui.cfg in the app startup area.
This is far from ideal because:
- it is another file to read and save over a session
- the default_ui_config.cfg file has to be included in the distribution
- since it easy for the user to mess up a docking layout, there probably has to be a new View menu item for 'revert to default layout'
- it relies on the binary independent store format, which can't be edited by hand

However... it does work. The next step may be to reverse engineer all the calls done in reading one of these files to work out how to control it and use the dADL config file to store the information.

Next steps in the rewrite of the GUI:
- make the 'action_bar' (currently an HBOX) into a proper SD_TOOL_BAR
- make the errors and console tabs in the status notebook separate tools