Reinstate the manifest file. We removed it long ago, in revision 115, because it wasn't really needed. As of EiffelStudio 6.8 it definitely wasn't needed (as explained by http://tech.groups.yahoo.com/group/eiffel_software/messages/18522?threaded=1&m=e&var=1&tidx=1 on the Eiffel Software mailing list). Starting with EiffelStudio 6.8, we got Windows visual styles automatically thanks to revision 86088:
--- Src/library/wel/Clib/disptchr.c     (revision 86087)
+++ Src/library/wel/Clib/disptchr.c     (revision 86088)
@@ -24,6 +24,10 @@
 #include "wel_globals.h"
 #include "wel.h"

+/* Pragma for automatically linking against Common Controls 6.0 */
+#pragma comment(linker,"\"/manifestdependency:type='win32' name='Microsoft.Windows.Common-Controls' version='6.0.0.0' processorArchitecture='*' publicKeyToken='6595b64144ccf1df' language='*'\"")
+

But revision 86090 made it apply only for Microsoft C:

+#ifdef _MSC_VER
+/* Pragma for automatically linking against Common Controls 6.0 */
+#pragma comment(linker,"\"/manifestdependency:type='win32' name='Microsoft.Windows.Common-Controls' version='6.0.0.0' processorArchitecture='*' publicKeyToken='6595b64144ccf1df' language='*'\"")
+#endif
+

Therefore, compiling with MingW didn't give us the Windows visual styles. Most notably, tooltips weren't working. One fix would be to deploy the manifest as a separate file, but it's simpler to reinstate the manifest and instruct the resource compiler to use it.

The manifest file that is included here was copied from the EIFGENs directory. It had been generated there during Microsoft C compilation on my computer, courtesy of the #pragma noted above.

I also had to modify the Eiffel SCons builder so that the resource compiler could find the manifest file in case the build was being done from a different directory. When the builder makes a temporary copy of the .rc file, it was substituting the path to the icon files; it now does likewise for the manifest.

This is Microsoft's explanation:
http://msdn.microsoft.com/en-us/library/windows/desktop/bb773175(v=vs.85).aspx

This is MingW's:
http://sourceforge.net/apps/mediawiki/libwdi/index.php?title=Reuse#Creating_a_Windows_GUI_application_that_use_the_latest_Windows_visual_enhancements_.28Aero_Glass_look_on_Vista_or_Windows_7.29.2C_even_if_compiled_from_MinGW.2Fcygwin

Are here are some thoughts from a SCons perspective (although we aren't using this):
http://www.scons.org/wiki/EmbedManifestIntoTarget